requirements.txt
alabaster==0.7.6
aniso8601==1.1.0
Babel==2.1.1
blinker==1.4
docopt==0.4.0
docutils==0.12
ecdsa==0.13
enum34==1.1.2
Flask==0.10.1
Flask-Babel==0.9
Flask-Cache==0.13.1
Flask-Classy==0.6.10
Flask-Cors==2.1.2
Flask-Login==0.3.2
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.4.2
Flask-WTF==0.12
gunicorn==19.4.5
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
mailchimp==2.0.9
MarkupSafe==0.23
marshmallow==2.4.2
marshmallow-jsonapi==0.4.2
MySQL-python==1.2.5
nose==1.3.7
paramiko==1.16.0
psycopg2==2.6.1
pycrypto==2.6.1
Pygments==2.0.2
PyJWT==1.4.0
python-dateutil==2.4.2
pytz==2015.7
redis==2.10.5
requests==2.8.1
sendgrid==1.5.18
shortuuid==0.4.2
six==1.10.0
smtpapi==0.3.1
snowballstemmer==1.2.0
speaklater==1.3
Sphinx==1.3.2
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.9
traceback2==1.4.0
twilio==6.2.dev0
unittest2==1.1.0
Werkzeug==0.11.2
wheel==0.24.0
WTForms==2.0.2
xmltodict==0.9.2
